troca no nome dos arquivos funcoes_aux.py e requisicoes.py, além de ajuste de documentação em algumas funções
